requirements.txt
numpy==1.14.5
opencv-python==3.4.1.15
pillow==5.1.0